Add files via upload
The Application starts off with the Login/Sign up page and then comes the user category page where the user selects the difficulty level, the category in which he/she are willing to take the quiz on, Number of questions, Type of answer mode(MCQ's/T-F) then the instructions page which shows the instructions that the attendee should be following in order to take the quiz which is designed with some additional hover and animated effects. Upon completing all these properly user will be taken to questions page where in each correct answer will be given +1 and wrong answer being 0. Finally the score and verdict will be displayed.This is how the application is built using all thw MERN Technologies.